refactor burner wallet into `useBurnerWallet`
